Changing the structure of storing the data, as startAT() / endAt() / equalTo() could only use when the child is within that ref ("users/") like users: {     "2019/01/19" : {          submitted: false,     } }
cannot use in this format:

users: {
    "2019/01/19" : {
        "01:20" : {
            submitted: false,
        }
    }
}

Those functions cannot be used when they are one more key in it, if you need, add child("01:20") after the ref in order to work. In this case, we do not know what the key in it, hence we need to restructure by combine both date and time together.