Changing the structure of the database.
Price is change to Amount.

Adding the email feature which will trigger the user to
choose the email application and attach the excel file to the email.

Furthermore, adding the feature of generate workbook using the data selected by the user in the submission page.

Note: Each month claim summary will be at their individual worksheet.